Moved idtype related stuff to idtype package
