+++ phovea_server/idtype/__init__.py
@@ -1,0 +1,6 @@
+from ..plugin import lookup
+from .mapper import MappingManager
+
+
+def get_mappingmanager() -> MappingManager:
+    return lookup('mappingmanager')
